Coveralls to report via webhook
This enables coveralls webhook, so that the actual coveralls results
appear on the pull request.
Also created a dedicated job for Docker testing so it can run in
parallel with the other tests.
Prior this change the CI was running in 3 minutes, it now runs in 2
minutes as the Docker job runs in parallel with the unit tests one.